+++ worker/trace/package_test.go
@@ -76,7 +76,7 @@ func (s *baseSuite) expectClient() {
 	s.client.EXPECT().Start(gomock.Any()).AnyTimes()
 	s.client.EXPECT().Stop(gomock.Any()).AnyTimes()
 
-	s.clientTracer.EXPECT().Start(gomock.Any(), gomock.Any(), gomock.Any()).DoAndReturn(func(ctx context.Context, name string, opts ...trace.SpanStartOption) (context.Context, trace.Span) {
+	s.clientTracer.EXPECT().Start(gomock.Any(), gomock.Any(), gomock.Any()).DoAndReturn(func(ctx context.Context, name string, opts ...trace.SpanStartOption) (context.Context, ClientSpan) {
 		return ctx, s.span
 	}).AnyTimes()
